requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
tqdm==4.66.1
matplotlib==3.7.2
tensorboard==2.14.0
pydantic==2.0.3
pyyaml==6.0
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.28.0
